libstagefright: Fix HDR profile mapping
Extractor was wrongly marking clips with color-transfer set to
HLG as HDR10 profiles.
Clips with color-transfer set to HLG will no longer be identified
as HDR10 clips now.

When the profile is a 10-bit profile, based on color-transfer
and hdr metadata, profiles are mapped as follows

(color-transfer set to ST2084 (PQ) or not present) and hdr10-plus-info
 maps to HDR10+ profiles
(color-transfer set to ST2084 (PQ) or not present) and hdr-static-info
 maps to HDR10 profiles

Bug: 231520574
Test: Verify profile of 10-bit clips with transfer set to HLG by using
 extractor
Test: atest android.mediav2.cts.ExtractorTest CtsMediaExtractorTestCases

Change-Id: Ia4a901c319f5e0c7361d0ba921dec9db5d9ca2ae
Signed-off-by: Dmitrii <bankersenator@gmail.com>