VP9: Synthesize CSD from frame header
If the CodecPrivate data is missing from the container, use the first
frame's frame header to parse and synthesize the CSD.

This is an update of c9d8dee4cb0f0fa6f41e7d8c324221f897f8b51f
which was reverted. This commit updates the above with support for
cases where the first frame is not a key frame.

Test: atest CtsMediaExtractorTestCases
Test: atest CtsMediaV2TestCases:ExtractorTest
Test: atest CtsMediaV2TestCases:ExtractorUnitTest
Test: atest CtsMediaV2TestCases:DecoderColorAspectsTest
Test: atest CtsMediaV2TestCases:EncoderProfileLevelTest
Test: atest CtsMediaDecoderTestCases:DecoderConformanceTest
Bug: 295804596
Google: 3032528
Change-Id: I4fd99dcbd736da368c68d802cc2212d4482bc2f0
Signed-off-by: Dmitrii <bankersenator@gmail.com>