issue: After connecting the Bluetooth earphones and projecting the screen, music will play from the earphones. At this point, disconnect the earphones, click play, and the projection will be silent
Analysis: The screen projection device is remote_stubmix, and the output: deepubuffer does not support this device. If the Bluetooth earphones in the testing steps are replaced with wired earphones, restoreTrack_1 will be called after disconnecting the earphones because hasCurrentEncodedFFormat returns true. When the Bluetooth earphones are disconnected, the mCurrentEncodedFFormat is set to AUDIO-FORMAT-DEFAULT in setDeviceConnectionStateInt, so the return value of hasCurrentEncodedFFormat remains false

Solution: Add a true judgment when Bluetooth earphones are disconnected in hasCurrentEncodedFFormat

issue ID: 405883200

Google: 3598871
Change-Id: I1e3237109cf5f59137e44ccafc706dc50894742c
Signed-off-by: fengxiaoshuang <fengxiaoshuang@xiaomi.corp-partner.google.com>
Signed-off-by: Dmitrii <bankersenator@gmail.com>